% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_difference.R
\name{intensity.difference}
\alias{intensity.difference}
\title{Intensity difference test}
\usage{
intensity.difference(values.1, values.2)
}
\arguments{
\item{values.1}{A numeric vector of values}

\item{values.2}{A numeric vector of values}
}
\description{
This function calculates an intensity corrected z-score
statistic for a matched pair of numeric vectors. For each
point a local sub-sampling of the data is performed where
a set of points with the most similar mean values across
the two datasets are collected.  The distribution of
differences for these selected points is then modelled with
a normal distribution and a z-score value for the point
being tested is then calculated, along with the p.value
corresponding to that z-score.  The absolute difference
and FDR corrected p-values are also reported.
}
\examples{
intensity.difference()
}
