% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_diff_plot.R
\name{intensity.diff.plot}
\alias{intensity.diff.plot}
\title{Intensity difference plot}
\usage{
intensity.diff.plot(intensity.diff.data)
}
\arguments{
\item{values.1}{A numeric vector of values}

\item{values.2}{A numeric vector of values}

\item{window.proportion}{What proportion of the data will be used to create each local model - Default 0.01}
}
\description{
This function takes the output from the intensity_difference
function and draws a number of diagnostic plots to help to 
evaluate the results of the analysis
}
\examples{
intensity_diff_plot()
}
